libs/utils/analysis: add CPUs module
This patch move the CPUs signals analysis code, originally part of the
trace_analysis.py module, to be a new TraceAnalysis module with dynamic
registration via the AnalysisRegister.
