+++ libs/utils/trace_analysis.py
@@ -142,83 +142,3 @@ class TraceAnalysis(object):
                 [df.cpu.isin(self.platform['clusters']['little'])],
                 [tip_lcap], tip_bcap)
 
-    def __plotCPU(self, cpus=None, label=''):
-        if cpus is None or len(cpus) == 0:
-            return
-        if label != '':
-            label1 = '{} '.format(label)
-            label2 = '_{}s'.format(label.lower())
-
-        # Plot required CPUs
-        fig, pltaxes = plt.subplots(len(cpus), 1, figsize=(16, 3*(len(cpus))));
-        plt.suptitle("{}CPUs Signals".format(label1),
-                     y=.99, fontsize=16, horizontalalignment='center');
-
-        idx = 0
-        for cpu in cpus:
-
-            # Reference axes to be used
-            axes = pltaxes
-            if (len(cpus) > 1):
-                axes = pltaxes[idx]
-
-            # Add CPU utilization
-            axes.set_title('{0:s}CPU [{1:d}]'.format(label1, cpu));
-            df = self.trace.df('sched_load_avg_cpu')
-            df = df[df.cpu == cpu]
-            if len(df):
-                df[['util_avg']].plot(ax=axes, drawstyle='steps-post', alpha=0.4);
-
-            # if self.trace.hasEvents('sched_boost_cpu'):
-            #     df = self.trace.df('sched_boost_cpu')
-            #     df = df[df.cpu == cpu]
-            #     if len(df):
-            #         df[['usage', 'boosted_usage']].plot(
-            #             ax=axes,
-            #             style=['m-', 'r-'],
-            #             drawstyle='steps-post');
-
-            # Add Capacities data if avilable
-            if self.trace.hasEvents('cpu_capacity'):
-                df = self.trace.df('cpu_capacity')
-                df = df[df.cpu == cpu]
-                if len(df):
-                    # data = df[['capacity', 'tip_capacity', 'max_capacity']]
-                    # data.plot(ax=axes, style=['m', 'y', 'r'],
-                    data = df[['capacity', 'tip_capacity' ]]
-                    data.plot(ax=axes, style=['m', '--y' ],
-                              drawstyle='steps-post')
-
-            axes.set_ylim(0, 1100);
-            axes.set_xlim(self.x_min, self.x_max);
-
-            # Disable x-axis timestamp for top-most cpus
-            if (len(cpus) > 1 and idx < len(cpus)-1):
-                axes.set_xticklabels([])
-                axes.set_xlabel('')
-            axes.grid(True);
-
-            idx+=1
-
-        # Save generated plots into datadir
-        figname = '{}/{}cpus{}.png'.format(self.plotsdir, self.prefix, label2)
-        pl.savefig(figname, bbox_inches='tight')
-
-    def plotCPU(self, cpus=None):
-        if not self.trace.hasEvents('sched_load_avg_cpu'):
-            logging.warn('Events [sched_load_avg_cpu] not found, '\
-                    'plot DISABLED!')
-            return
-
-        # Filter on specified cpus
-        if cpus is None:
-            cpus = sorted(self.platform['clusters']['little'] + self.platform['clusters']['big'])
-
-        # Plot: big CPUs
-        bcpus = set(cpus) & set(self.platform['clusters']['big'])
-        self.__plotCPU(bcpus, "big")
-
-        # Plot: LITTLE CPUs
-        lcpus = set(cpus) & set(self.platform['clusters']['little'])
-        self.__plotCPU(lcpus, "LITTLE")
-
